organize the pill bars into another folder
